SWWEB-478 FINAL
- Made the application name being used in the MaxCommReadFlag table
dynamic in both the commlog_service.js which writes the data and the
OtbServiceRequestDataSet.cs which reads the data.
- Added a parameter to the doToggle function which will allow for the
original composition list data to be sent to the doToggle function. This
allowed us to remove the iterative lookup from the commlog_service.js
and allowed for us to properly check if the read flag is true or false.
This was occurring because we primarily send the detail data to the
doToggle function, and only once in a while do we fetch the original
list data. For this unread/read functionality, we need to have the
original composition list data every single time.