scanfilter modal fix
The scan filter modal's columns resize when rows are added depending on
the length of the field name being added. Add css to the site.css to fix
the widths.

The add row will now only display when there are items that can be added
to the list of filter fields.