Inventory Return/Issue & Additional Features
- Qty Returned correctly displays on invissue detail/list (This required
some custom queries in the OtbInventoryIssueDataSet)
- Return buttons correctly display
- User is navigated to correct detail upon clicking a matusetrans record
in the invissue grid
- Returns are created successfully for existing issues
- Custom formatter services can be displayed on a renderer parameter
using params="formatter=@service.method;"
- max value of a numeric input can be setup as a calculation of two
fields, i.e. params="min=1;max=@quantity - @qtyreturned;" However, these
fields are not yet being watched so updates to the datamap will not
affect the max counter
- Refactored eventdispatcher_service into an event_service and
dispatcher_service. This was needed to reuse code for the dynamic
injection of a service for the formatter.