Correct reference to local function
Inventory service was trying reference a function defined in its return.
Moved the function so that it is defined on the service making it so
that it can be referenced by the functions it passes back. Also
corrected the metadata so that changing the item would call the
afterchangeinvissueitem method.