Inventory - Batch Issuing
- Added update functionality to existing items on the batch issue
composition list. Currently the showexpression attribute is not working
on commands, so both the add and update buttons are showing whenever you
try to add or edit an item to the batch
- Added a generic delete button action for compositionlists. This
resource can be added as a section, and will display a trash icon that
will simply delete the compositionitem from the clonedcompositiondata
(thus, removing it from your screen). This does NOT update Maximo