SWWEB-717
- Created two qualified relationships from invissue (matusetrans) entity
to invbalances entity. Two relationships are needed since we're using
the same entity to look up bins and lots and need to use the
reverseLookupAttribute function to specify the binnum (right now, the
Bin is being set with the lotnum because the relationship has the lotnum
as the primary and the lotnum is required for the primary in order to
get all instances of "binnum/lotnum".
- Added lotnum to the inventory issue/return list
- Sectioned off the individual new issuing detail page to be similarly
sectioned off to the new item batch issuing page
- Renamed 'Issue To' label to 'Issued To' label and moved the Issued To
field towards the top of the new inventory issuing page.
- Created two different schemas to the invbalances application. One for
the bin lookup modal, and one for the lot lookup modal.
- Added ability to add a DefaultStoreloc property for the property.xml
metadata. This is used similarly to orgid and siteid.
- Added storeloc to the SWDB sw_user2 table. This will store the user's
storeroom.
- Updated user setup pages to include a user's storeroom (adding/editing
a user), as well as using the Default storeroom property when creating
the user's profile.
- Added the autopopulation of new inventory records to automatically be
set to the user's current storeroom
- Fixed bug which was not correctly transferring/issuing from inventory
balances that didn't have a bin. In this case, because of having a
"null" value for the bin, the inventory record's default binnum was
being used. I changed this to set the binnum to an empty string ("") in
the case that it was set to null. This required the creation of a custom
creation button for the new inventory issue page.
- Whenever an item is selected, the binnum will populate with the
inventory record's default binnum.
- Added filters for lots/bins to the OtbInventoryIssueDataSet.cs