SWWEB-717 Changes
- Added custom cancel button redirect to issues/returns list for the
detail pages which had errors using the framework's cancel button
- Added issuetype to newInvIssueDetail -- this was a bug preventing new
issue records from saving
- Added a few fields to both newitem and newinvissuedetail that will
check for afterchanges on the lot and bin inventory balance's
currentbalance, binnum, and lotnum fields. This was needed because both
targets (binnum and lotnum) had afterchange events that would change
each other's value. This would cause a lot of issues with incorrect data
and in some cases, infinite loops. Instead, by having the after change
on the extraprojectionvalues i can determine which field has been
changed and therefore the correct currentbalance, lotnum, and binnum to
display.
- Updated filter bins to the new inventory balance structures. Now the
"empty" bin is showing again.
- Updated method to hide the "new item" popup modal when trying to add
an item in the batch issuing page. I am now ensuring that the data is
wiped/cleaned before hiding the modal. However, because of an issue with
lookup fields, the values still 'appear' to be in each lookup.
- Added a few fields to the itemlookup schema (fetching related
inventory information)
- Added default storeroom to the user's 'My Profile' page. This was
modeled after the orgid/siteid.
- Changed onload event call from the compositionlist to use parentdata
instead of previousdata. It seems that the data available has changed,
and luckily, no one else is using this onload parameter.
- Updated 'additemtobatch' method in the inventory_service.js. This
method is called whenever an item is scanned. Now, the correct fields
will be fetched (inventory cost, default binnum, etc). so that the
records can submit. However, there is an issue where the items are
trying to submit too quickly back to back. I will investigate a fix for
the issue this week. Additionally, I need to add one more lookup before
the scanned item is added to the list of items: the default bin's lotnum
(if there is one). Additionally, if the item is a tool, I need to
display the popup modal to update the item and an alert notifying them
that the 'Issued To' field must be filled out if the itemtype is TOOL.
- Added a check for all new issue pages (Batch and individual) that will
warn the user with an alert if the itemtype is TOOL and the Issued To
field is not filled out.
- Added a check on the new issue pages (Batch and Invidual) that will
warn the user with an alert if the item is not filled out.
- After setting the item in a new inventory issue page (Batch and
Individual), the lotnum/binbalance_.curbal field will be set to null.
The curbal change will trickle down with another afterchange event that
will update the #curbal attribute shown on the screen.
- Fixed a bug with the submit button of the newInvIssue page (individual
issuing). This was redirecting to the wrong application schema, which
would take the user nowhere.
- Prepared afterchange events so that the binnum/lotnum lookups will
affect each other (i.e. update binnum with associated lotnum, update
lotnum with associated binnum), however, this will not work until the we
fix the bug mentioned above where the lookup values 'look' like they
havent changed.