SWWEB-717
- Modified 'lookupattribute' to instead be named 'forcedistinctoptions'
attribute. When forceddistinctoptions is false, then duplicate labels
(field which is primary on relationship) can be added to the
resultobject.
- Modified primary key on binnum's relationship to invbalances so that
the primary key is now 'binnum' (rather than using lotnum to force
"duplicate" binnums to show)
- Fixed an issue where an angular error was showing up in chrome if the
itemnum was cleared. This was due to the
doUpdateUnitCostFromInventoryCost function executing even when the
itemnum is null.
- Added the itemtype to the item lookup so that we can prepopulate the
new matusetrans record after scanning an item to issue (bulk issuing)
- Added some code for when a tool is scanned. This is currently
commented and will need to be improved.