SWWEB-717
- Put in a workaround to prevent 5 lookup modals from displaying on new
inv issue detail page. This was caused due to the lookup being outside
of a section while there are other lookups existing in other sections.
- Added a "lookuponblur" renderer parameter for lookups that will
determine whether or not the lookup will display after blurring a lookup
field.
- Added a lookuponblur=false parameter to the transfer inventory
application's "to bin". This is because you can enter in a free-text bin
(i.e. one that doesnt necessarily exist), however you can still search
for a bin if you would like to use the lookup.
- Renamed required to requiredexpression (next commit will implement the
expressionservice call)
- Fixed showexpression for commands. Now, if the showexpression is false
for a command, it will not be displayed.