Add comments and new doc folder to sW4.Web
Added some comments to the scanner detection service for
initInventoryGridListener to describe where it is pulling the scan order
from.
Added a new doc folder to the softWrench.sW4.Web project, as advised by
Luiz, and added a short documentation file explaining what the scanner
detection service is doing and describing one of the functions it has as
an example.