SWWEB-858
- Added three new event types for event schemas:
-- beforesubmit.prevalidation
-- beforesubmit.onvalidation
-- beforesubmit.postvalidation
- beforesubmit.prevalidation & beforesubmit.postvalidation event types:
Works as a normal custom service method where we can calculate/transform
data, however, there is support to call a confirmation alert in either
the prevalidation or postvalidation method. The only requirement to
using the event as a confirmation is that you must return false from the
custom service (or return alert.confirm, which will return a false by
default), and in the success function you will need to call
parameters.continie();
EX:
return alertService.confirm(null, null, function () {
parameters.continue();
}, "Are you sure you want to accept?");
- beforesubmit.onvalidation Event: Attaches to a custom service/method
reference and appends the array of strings to the red validation error
banner. This occurs during the creation or update of a record.
- Removed the custom submit button from the returnInvIssuedetail page.
Now, the framework is handling this through the prevalidation and
postvalidation events.
- Updated event_service.js to support beforesubmit.onvalidation,
beforesubmit.prevalidation, and beforesubmit.postvalidation event types.
- Added a return false at the end of the confirm method in
alert_service.js so that we can return the alertservice.confirm in our
prevalidation / postvalidation event types, thus halting the
validation/submit calls in crud_body until the parameters.continue
method is called from the alert_service's "OK" confirmation click.
- Added a check on the messagehandler.js's $on sw_validationerrors event
handler. This will check that the validationArray is not null and
contains at least 1 element before displaying the validation bar.
- Modified crud_body's submit service so that there are functions to
call continue with in the case that the program's flow is interrupted by
an alert_service call in the beforesubmit.prevalidation or
beforesubmit.postvalidation's custom method.
- The datamap is now copied using angular.copy() in the crud_body's save
method. This is to ensure that the screen values are not changed in the
case that the submit fails. Even more so important now that we have
custom beforesubmit functions that can manipulate the datamap.
- Updated schemas that used the "oncrudsaveevent.validationservice"
property to instead use the beforesubmit.onvalidation event.