Correct bug in search service
There was a bug in the search service causing the execution of the JS to
break. Also updated the scanner detection service so that if a scanned
value returns either zero rows, or multiple rows of possible data, no
value will be set and the user will be warned that the scan could not be
used to set the value.