SWWEB-826
- $scope.originalDatamap is copied whenever the crud_body registered a
sw_bodyrenderevent. This will be used to "clean" the datamap before
sending it to the MIF
- Before modifying the datamap in the crud_body's save function, we make
a copy of the datamap using angular.copy so that we preserve its values
in the case of a submission failure. This way, the same values show on
the screen to the user and we can reapply the
beforesubmit/prevalidation,postvalidation transformations to the prior
state of the datamap.
- Added a handleDatamapForMIF function in the submit_service which will
compare the original datamap to the current datamap and update any
"removed" values (i.e. originally were not null, but are now null) with
an empty string " ". The cleaning process is applied only to Field,
AssociationField, and OptionField renderers. However, a section's
displayables will be cleaned as well.
- Added a call to the handleDatamapForMIF function in the crud_body's
save function.
- Removed the functionality from the lookupCodeChange function
(crud_input_fields) that would set the attribute's datamap/associated
values to null.
- Moved updateAssociations function from scannerdetection_service and
moved it to the associationService. I renamed this to
updateDependentAssociationValues since it will update the dependent
association values for eager associations / non eager associations
(compared to the function already named 'updateAssociations' in the
associationService).
- Reused the updateDependentAssociationValues function on the
lookupCodeBlur function. Whenever the value in the lookup is NOT NULL,
and the value does not match what is currently in the datamap, the
updateDependentAssociationValues will be executed on the lookup's new
value. After, the lookupModal is displayed. However, after my next merge
with develop I will be changing this to check whether or not multiple
results were found during the updateDependentAssociationValues function.
If there are multiple results, then we will be showing the lookup modal.