SWWEB-826
- Added a "after fetch" hook to the updateDependentAssociationValues
function in associationService.js. This hook can be used to perform
custom validation (as is being used in the scannerdetection_service) or
to perform a whole other set of functionality (as is now being performed
when there are multiple association options after typing in a new
association code). By returning a false, the final code won't be
executed to update the lookupAssociationCode,
lookupAssociationDescription, and
updateAssociationOptionsRetrievedFromServer.
- Added selectedOption to the lookupModal directive which is set with
the option that is selected. This will keep track of when an option has
been selected rather than an 'exit' or a 'close' command.
- Added a modalCanceled variable to the scope to keep track of when the
modal was canceled.
- After the association is canceled or exited, the hide method will now
clear the associationCode, Description and update the underlying
association objects w/ null.
- Added a 'handleMultipleLookupOptionsFn' to the crud_input_fields.js
that will serve as the custom hook during the
updateDependentAssociationValues function call in the lookupCodeBlur
function. If their is only one result, the
updateDependentAssociationValues function finishes -- updating the
association values with the one result. If there is not a single result,
then the lookup modal is displayed using the association code as the
search parameter.