Update inv issue when selecting rotating asset
After selecting the rotating asset the bin and curbal should be updated
to the binnum of the rotating asset selected and a curbal of one.

There was an issue with lookups not normalizing the case of the string
values before comparing them to other string values from from the server
causing some items not to be found on the initial lookup.