Reported by now saves when creating SR
Reported by was not on the new SR schema so it was never saved to the
record.
Also updated the location and asset for SR to be dependant on the
siteid.
Updated the location field for SR and WO to use a lookup instead of the
default auto complete drop down.