GRICC-111
After comparing the previous value to the value retreived form the
server, set the datamap to the server value so that it matches up with
the information in the association options. This avoids issues caused
when users enter values in lower case that are stored in the database in
uppercase.