Update labor relationship for laborcode scanning
HAd ot make the laborcode the primary attribute of the relationship
between matusetrans and labor so that when querying the server, the
labor code would be matched correctly with the labor code field. Also
removed the itemnum's dependency on the store room to allow the default
itemnum to fill when the store room was blank.