Hack to correct changedate issue
Change date on records is being set based on the utc from the
application, our query checks to see if the datetime of the record is <
the current date time. This is causing records that are submitted to not
show up for an hour. Chnaged the query to allow for records up to two
hours in the future to display.