Two Fixes for Bin Lookup
- Changed inventory service so that it only registers a "null" bin as a
removal of the bin value. Before, it was checking for null or empty. Now
that there are association options for bins that have an empty string
"", we cannot treat the empty string as a "removal" of the bin value.
- Changed inventory service so that the current balance / lotnum are
removed if the binnum is null, regardless of whether or not binbalances_
data is available.