Update composition ng-repeat to remove ng-init
Previously the composition lists ng-init would set a rowindex var to be
used for getting data from lists for the rows. This worked when not
removing rows, but since it is not updated after the initial init,
removing a row would cause the index's to be incorrect. Updated to use a
track by on the ng-repeat which is updated when rows are removed.