Add support for systemid to location application
When selecting a location from the grid, the systemid of the row
selected needs to propogate through the system so that it is available
when the detail page is shown. This allows us to show two separate
detail pages for a single location, one for each system the location is
a part of.